v23.1.42 is released

+++ controls/grids/src/grid/renderer/content-renderer.ts
@@ -287,7 +287,8 @@ export class ContentRender implements IRenderer {
     public refreshContentRows(args: NotifyArgs = {}): void {
         const gObj: IGrid = this.parent;
         if (gObj.currentViewData.length === 0) { return; }
-        if (gObj.editSettings.mode === 'Batch' && gObj.editModule.formObj && gObj.editSettings.showConfirmDialog === false) {
+        if (gObj.editSettings && gObj.editModule && gObj.editSettings.mode === 'Batch' && gObj.editModule.formObj
+            && gObj.editSettings.showConfirmDialog === false) {
             gObj.editModule.destroyForm();
         }
         const dataSource: Object = this.currentMovableRows || gObj.currentViewData;
